rename: MenuItemIconProps.ts→IconProps.ts
